requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
chromadb
sentence-transformers
pdfplumber
python-dotenv
pydantic
